File operation
Removed get_age from users.py and added it to baseuser.py

Removed BaseUser and PartialUser from gamepass.py

Rmoved inventory.py
