+++ ro_py/gamepass.py
@@ -9,11 +9,9 @@ import iso8601
 import asyncio
 
 from ro_py.events import Event
-from ro_py.users import BaseUser
 from typing import Tuple, Callable
 from ro_py.events import EventTypes
 from ro_py.utilities.errors import NotFound
-from ro_py.bases.baseuser import PartialUser
 from ro_py.utilities.pages import Pages, SortOrder
 from ro_py.utilities.clientobject import ClientObject
