+++ ro_py/users.py
@@ -61,24 +61,6 @@ class User(BaseUser, ClientObject):
         # has_premium_req = requests.get(f"https://premiumfeatures.roblox.com/v1/users/{self.id}/validate-membership")
         # self.has_premium = has_premium_req
         
-        
-    async def get_age(self):
-        """
-        Gets a users account age.
-        :return: int
-        """
-        user_info_eq = await self.requests.get(endpoint + f"v1/users/{self.id}")
-        user_info = user_info_req.json()
-        user = user_info['name']
-        
-        year = int(user.created.strftime("%Y")
-        month = int(user.created.strftime("%m"))
-        day = int(user.created.strftime("%d"))
-        now = date.today()
-       years = now.year - year
-       months = now.month - month
-       days = now.day - day + years * 365 + months * 31
-       return days
